% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{actions.Board}
\alias{actions.Board}
\title{Conseguir acciones para un problema de Board}
\usage{
\method{actions}{Board}(problem, state)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado actual del problema}
}
\value{
Posibles acciones
}
\description{
Conseguir acciones para un problema de Board
}
