% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiBoard.R
\name{actions.MultiBoard}
\alias{actions.MultiBoard}
\title{Conseguir acciones para un problema de MultiBoard}
\usage{
\method{actions}{MultiBoard}(problem, state)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado actual del problema}
}
\value{
Posibles acciones
}
\description{
Conseguir acciones para un problema de MultiBoard
}
