% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-queens.R
\name{actions.NQueens}
\alias{actions.NQueens}
\title{Conseguir acciones para un problema de n-reinas}
\usage{
\method{actions}{NQueens}(problem, state)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado actual del problema}
}
\value{
Posibles acciones
}
\description{
Conseguir acciones para un problema de n-reinas
}
