% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astar.R
\name{astar}
\alias{astar}
\title{A* algorithm}
\usage{
astar(problem, heuristic.method = "ucs")
}
\arguments{
\item{problem}{Problema a buscar}

\item{heuristic}{Función heurística}
}
\description{
A* algorithm
}
