% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{checkValidState.Board}
\alias{checkValidState.Board}
\title{Check State for Board Problem}
\usage{
\method{checkValidState}{Board}(problem, state, action)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado actual}

\item{action}{Accion tomada}
}
\description{
Check State for Board Problem
}
