% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiBoard.R
\name{checkValidState.MultiBoard}
\alias{checkValidState.MultiBoard}
\title{Comprobar validez de un estado de un problema de MultiBoard}
\usage{
\method{checkValidState}{MultiBoard}(problem, state, action)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado actual}

\item{action}{Última accion tomada}
}
\value{
Estado corregido para ser válida
}
\description{
Comprobar validez de un estado de un problema de MultiBoard
}
