% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{checkValidState}
\alias{checkValidState}
\title{checkValidState}
\usage{
checkValidState(problem, state, action)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado actual}

\item{action}{Accion tomada}
}
\description{
checkValidState
}
