% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{costState.Board}
\alias{costState.Board}
\title{costState for Board Problem}
\usage{
\method{costState}{Board}(problem, state)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado a calcular el coste}
}
\description{
costState for Board Problem
}
