% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiBoard.R
\name{costState.MultiBoard}
\alias{costState.MultiBoard}
\title{Calcular coste para un estado de un problema de MultiBoard}
\usage{
\method{costState}{MultiBoard}(problem, state)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado del que se va a calcular el coste}
}
\value{
Coste del estado dado
}
\description{
Calcular coste para un estado de un problema de MultiBoard
}
