% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-queens.R
\name{costState.NQueens}
\alias{costState.NQueens}
\title{Calcular coste para un estado de un problema de n-reinas}
\usage{
\method{costState}{NQueens}(problem, state)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado del que se va a calcular el coste}
}
\value{
Coste del estado dado
}
\description{
Calcular coste para un estado de un problema de n-reinas
}
