% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{finalState.Board}
\alias{finalState.Board}
\title{Final State for Board Problem}
\usage{
\method{finalState}{Board}(problem, state)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado a comparar si es final}
}
\description{
Final State for Board Problem
}
