% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiBoard.R
\name{finalState.MultiBoard}
\alias{finalState.MultiBoard}
\title{Calcular si un estado es final para un problema MultiBoard dado}
\usage{
\method{finalState}{MultiBoard}(problem, state)
}
\arguments{
\item{problem}{Problema a comprobar}

\item{state}{Estado a comparar si es final}
}
\value{
Comprobación del estado
}
\description{
Calcular si un estado es final para un problema MultiBoard dado
}
