% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-queens.R
\name{finalState.NQueens}
\alias{finalState.NQueens}
\title{Calcular si un estado es final para un problema n-reinas dado}
\usage{
\method{finalState}{NQueens}(problem, state)
}
\arguments{
\item{problem}{Problema a comprobar}

\item{state}{Estado a comparar si es final}
}
\value{
Comprobación del estado
}
\description{
Calcular si un estado es final para un problema n-reinas dado
}
