% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{finalState}
\alias{finalState}
\title{finalState}
\usage{
finalState(problem, state)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado a comparar si es final}
}
\description{
finalState
}
