% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiBoard.R
\name{getMultiBoard}
\alias{getMultiBoard}
\title{Construir MultiBoard}
\usage{
getMultiBoard(n = 20)
}
\arguments{
\item{n}{Longitud del tablero cuadrado}
}
\value{
Problema MultiBoard
}
\description{
Construir MultiBoard
}
