% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-queens.R
\name{getNQueens}
\alias{getNQueens}
\title{Construir Problema n-reinas}
\usage{
getNQueens(n = 8)
}
\arguments{
\item{n}{Longitud Problema N-reinas}
}
\value{
Problema N-Reinas
}
\description{
Construir Problema n-reinas
}
