% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{heuristic.Board}
\alias{heuristic.Board}
\title{Heuristic for Board Problem}
\usage{
\method{heuristic}{Board}(problem, state, heuristic.method)
}
\arguments{
\item{problem}{Problema a calcular}

\item{state}{Estado}

\item{heuristic.method}{Método}
}
\description{
Heuristic for Board Problem
}
