% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiBoard.R
\name{heuristic.MultiBoard}
\alias{heuristic.MultiBoard}
\title{Método heurístico para problema de MultiBoard}
\usage{
\method{heuristic}{MultiBoard}(problem, state, heuristic.method)
}
\arguments{
\item{problem}{Problema a calcular}

\item{state}{Estado}

\item{heuristic.method}{Método}
}
\description{
Método heurístico para problema de MultiBoard
}
