% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-queens.R
\name{heuristic.NQueens}
\alias{heuristic.NQueens}
\title{Método heurístico para problema de n-reinas}
\usage{
\method{heuristic}{NQueens}(problem, state, heuristic.method)
}
\arguments{
\item{problem}{Problema a calcular}

\item{state}{Estado}

\item{heuristic.method}{Método}
}
\description{
Método heurístico para problema de n-reinas
}
