% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{heuristic}
\alias{heuristic}
\title{Heuristic for Board Problem}
\usage{
heuristic(problem, state, heuristic.method = "ucs")
}
\arguments{
\item{problem}{Problema a calcular}

\item{state}{Estado}

\item{heuristic.method}{Método}
}
\description{
Heuristic for Board Problem
}
