% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{initialState.Board}
\alias{initialState.Board}
\title{Initial State for Board Problem}
\usage{
\method{initialState}{Board}(problem)
}
\arguments{
\item{problem}{Problema para obtener el estado inicial}
}
\description{
Initial State for Board Problem
}
