% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiBoard.R
\name{initialState.MultiBoard}
\alias{initialState.MultiBoard}
\title{Estado inicial para un problema de MultiBoard}
\usage{
\method{initialState}{MultiBoard}(problem)
}
\arguments{
\item{problem}{Problema a resolver}
}
\value{
Estado inicial para el problema dado
}
\description{
Estado inicial para un problema de MultiBoard
}
