% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-queens.R
\name{initialState.NQueens}
\alias{initialState.NQueens}
\title{Estado inicial para un problema de n-reinas}
\usage{
\method{initialState}{NQueens}(problem)
}
\arguments{
\item{problem}{Problema a resolver}
}
\value{
Estado inicial para el problema dado
}
\description{
Estado inicial para un problema de n-reinas
}
