% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{nextState.Board}
\alias{nextState.Board}
\title{Calcular el siguiente estado para un problema de Board}
\usage{
\method{nextState}{Board}(problem, state, action)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado actual}

\item{action}{Accion tomada}
}
\value{
Nuevo estado
}
\description{
Calcular el siguiente estado para un problema de Board
}
