% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiBoard.R
\name{nextState.MultiBoard}
\alias{nextState.MultiBoard}
\title{Calcular el siguiente estado para un problema de MultiBoard}
\usage{
\method{nextState}{MultiBoard}(problem, state, action)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado actual}

\item{action}{Accion tomada}
}
\value{
Nuevo estado
}
\description{
Calcular el siguiente estado para un problema de MultiBoard
}
