% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-queens.R
\name{nextState.NQueens}
\alias{nextState.NQueens}
\title{Calcular el siguiente estado para un problema de n-reinas}
\usage{
\method{nextState}{NQueens}(problem, state, action)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado actual}

\item{action}{Accion tomada}
}
\value{
Nuevo estado
}
\description{
Calcular el siguiente estado para un problema de n-reinas
}
