% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{nextState}
\alias{nextState}
\title{nextState}
\usage{
nextState(problem, state, action)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado actual}

\item{action}{Accion tomada}
}
\description{
nextState
}
