% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{path}
\alias{path}
\title{Path}
\usage{
path(problem, state, closed)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado actual}

\item{closed}{Conjunto de estados cerrados sobre los que reconstruir el camino}
}
\description{
Path
}
