% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representPath_Board.R
\name{representPath.Board}
\alias{representPath.Board}
\title{Represent Solution for Board Problem}
\usage{
\method{representPath}{Board}(problem, path)
}
\arguments{
\item{problem}{Problema a representar}

\item{path}{Camino de la solución}
}
\description{
Represent Solution for Board Problem
}
