% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representPath_MultiBoard.R
\name{representPath.MultiBoard}
\alias{representPath.MultiBoard}
\title{Mostrar gráfico con solución de un problema MultiBoard}
\usage{
\method{representPath}{MultiBoard}(problem, path)
}
\arguments{
\item{problem}{Problema a representar}

\item{path}{Camino de la solución encontrada}
}
\description{
Mostrar gráfico con solución de un problema MultiBoard
}
