% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representPath_NQueens.R
\name{representPath.NQueens}
\alias{representPath.NQueens}
\title{Mostrar gráfico con solución de un problema n-reinas}
\usage{
\method{representPath}{NQueens}(problem, path)
}
\arguments{
\item{problem}{Problema a representar}

\item{path}{Camino de la solución encontrada}
}
\description{
Mostrar gráfico con solución de un problema n-reinas
}
