% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-queens.R
\name{sameState.NQueensState}
\alias{sameState.NQueensState}
\title{Comparación de la igualdad de dos estados de n-reinas}
\usage{
\method{sameState}{NQueensState}(state1, state2)
}
\arguments{
\item{state1}{Estado 1 a comparar}

\item{state2}{Estado 2 a comparar}
}
\value{
Comprobación de la igualdad de dos estados
}
\description{
Comparación de la igualdad de dos estados de n-reinas
}
