% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{searchList}
\alias{searchList}
\title{searchList}
\usage{
searchList(state, search.list)
}
\arguments{
\item{state}{Estado a buscar}

\item{search.list}{Lista sobre la que buscar el estado}
}
\description{
searchList
}
