% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{setCost.Board}
\alias{setCost.Board}
\title{Establecer el coste para la reconstrucción de un problema de Board}
\usage{
\method{setCost}{Board}(problem, state, new.cost)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado del que se cambia el coste}

\item{new.cost}{Coste calculado}
}
\value{
Problema con el coste cambiado
}
\description{
Establecer el coste para la reconstrucción de un problema de Board
}
