% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{setCost}
\alias{setCost}
\title{setCost}
\usage{
setCost(problem, state, new.cost)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado a cambiar el coste}

\item{new.cost}{Nuevo coste a establecer}
}
\description{
setCost
}
