% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudoku.R
\name{writeSudoku}
\alias{writeSudoku}
\title{Write Sudoku}
\usage{
writeSudoku(board, file = "sudoku.sud")
}
\arguments{
\item{board}{Board to be written in .sud format}

\item{file}{File name to write sudoku. Default is sudoku.sud}
}
\description{
Write Sudoku
}
