% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{actions.Board}
\alias{actions.Board}
\title{Get actions for Board Problem}
\usage{
\method{actions}{Board}(problem, state)
}
\arguments{
\item{problem}{Board problem}

\item{state}{Estado actual del problema}
}
\description{
Get actions for Board Problem
}
