% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{checkValidState.Board}
\alias{checkValidState.Board}
\title{Comprobar validez de un estado de un problema de Board}
\usage{
\method{checkValidState}{Board}(problem, state, action)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado actual}

\item{action}{Última accion tomada}
}
\value{
Estado corregido para ser válida
}
\description{
Comprobar validez de un estado de un problema de Board
}
