% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{costState.Board}
\alias{costState.Board}
\title{Calcular coste para un estado de un problema de Board}
\usage{
\method{costState}{Board}(problem, state)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado del que se va a calcular el coste}
}
\value{
Coste del estado dado
}
\description{
Calcular coste para un estado de un problema de Board
}
