% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{finalState.Board}
\alias{finalState.Board}
\title{Calcular si un estado es final para un problema Board dado}
\usage{
\method{finalState}{Board}(problem, state)
}
\arguments{
\item{problem}{Problema a comprobar}

\item{state}{Estado a comparar si es final}
}
\value{
Comprobación del estado
}
\description{
Calcular si un estado es final para un problema Board dado
}
