% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{heuristic.Board}
\alias{heuristic.Board}
\title{Método heurístico para problema de Board}
\usage{
\method{heuristic}{Board}(problem, state, heuristic.method)
}
\arguments{
\item{problem}{Problema a calcular}

\item{state}{Estado}

\item{heuristic.method}{Método heurístico. Posibles acciones: euclidean, euclidean2, manhattan, chessboard, ucs}
}
\description{
Método heurístico para problema de Board
}
