% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{initialState.Board}
\alias{initialState.Board}
\title{Estado inicial para un problema de Board}
\usage{
\method{initialState}{Board}(problem)
}
\arguments{
\item{problem}{Problema a resolver}
}
\value{
Estado inicial para el problema dado
}
\description{
Estado inicial para un problema de Board
}
