% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{nextState.Board}
\alias{nextState.Board}
\title{Next State for Board Problem}
\usage{
\method{nextState}{Board}(problem, state, action)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado actual}

\item{action}{Accion tomada}
}
\description{
Next State for Board Problem
}
