% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representPath_Board.R
\name{representPath.Board}
\alias{representPath.Board}
\title{Mostrar gráfico con solución de un problema Board}
\usage{
\method{representPath}{Board}(problem, path)
}
\arguments{
\item{problem}{Problema a representar}

\item{path}{Camino de la solución encontrada}
}
\description{
Mostrar gráfico con solución de un problema Board
}
