% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{sameState.BoardState}
\alias{sameState.BoardState}
\title{Same State for Board Problem}
\usage{
\method{sameState}{BoardState}(state1, state2)
}
\arguments{
\item{state1}{Estado 1 a comparar}

\item{state2}{Estado 2 a comparar}
}
\description{
Same State for Board Problem
}
