% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{setCost.Board}
\alias{setCost.Board}
\title{setCost for Board Problem}
\usage{
\method{setCost}{Board}(problem, state, new.cost)
}
\arguments{
\item{problem}{Problema a resolver}

\item{state}{Estado a cambiar el coste}

\item{new.cost}{Nuevo coste a añadir}
}
\description{
setCost for Board Problem
}
